move cask/cmd/uninstall to cask/uninstall
