+++ Library/Homebrew/cask/cmd.rb
@@ -14,7 +14,6 @@ require "cask/cmd/audit"
 require "cask/cmd/fetch"
 require "cask/cmd/install"
 require "cask/cmd/reinstall"
-require "cask/cmd/uninstall"
 require "cask/cmd/upgrade"
 
 module Cask
