+++ Library/Homebrew/cmd/uninstall.rb
@@ -6,10 +6,10 @@ require "formula"
 require "diagnostic"
 require "migrator"
 require "cli/parser"
-require "cask/cmd"
 require "cask/cask_loader"
 require "cask/exceptions"
 require "cask/installer"
+require "cask/uninstall"
 require "uninstall"
 
 module Homebrew
@@ -76,7 +76,7 @@ module Homebrew
         Cask::Installer.new(cask, verbose: args.verbose?, force: args.force?).zap
       end
     else
-      T.unsafe(Cask::Cmd::Uninstall).uninstall_casks(
+      Cask::Uninstall.uninstall_casks(
         *casks,
         verbose: args.verbose?,
         force:   args.force?,
